+++ plutus-use-cases/plutus-use-cases.cabal
@@ -51,6 +51,10 @@ library
         Plutus.Contracts.Stablecoin
         Plutus.Contracts.Swap
         Plutus.Contracts.Uniswap
+        Plutus.Contracts.Uniswap.Data
+        Plutus.Contracts.Uniswap.OnChain
+        Plutus.Contracts.Uniswap.OffChain
+        Plutus.Contracts.Uniswap.Pool
         Plutus.Contracts.Vesting
     hs-source-dirs: src
     default-language: Haskell2010
