+++ plutus-use-cases/src/Plutus/Contracts/Uniswap/Data.hs
@@ -1,0 +1,127 @@
+{-# LANGUAGE DataKinds                  #-}
+{-# LANGUAGE DeriveAnyClass             #-}
+{-# LANGUAGE DeriveGeneric              #-}
+{-# LANGUAGE DerivingStrategies         #-}
+{-# LANGUAGE FlexibleContexts           #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE MultiParamTypeClasses      #-}
+{-# LANGUAGE NoImplicitPrelude          #-}
+{-# LANGUAGE OverloadedStrings          #-}
+{-# LANGUAGE RecordWildCards            #-}
+{-# LANGUAGE TemplateHaskell            #-}
+{-# LANGUAGE TypeFamilies               #-}
+{-# LANGUAGE TypeOperators              #-}
+{-# options_ghc -fno-warn-orphans          #-}
+{-# options_ghc -Wno-redundant-constraints #-}
+{-# options_ghc -fno-strictness            #-}
+{-# options_ghc -fno-specialise            #-}
+
+module Plutus.Contracts.Uniswap.Data
+  where
+
+import           Ledger
+import           Ledger.Value        (AssetClass (..), assetClass, assetClassValue, assetClassValueOf)
+import           Playground.Contract (FromJSON, Generic, ToJSON, ToSchema)
+import qualified PlutusTx
+import           PlutusTx.Prelude
+import qualified Prelude
+import           Text.Printf         (PrintfArg)
+
+-- | Uniswap coin token
+data U = U
+PlutusTx.makeIsDataIndexed ''U [('U, 0)]
+PlutusTx.makeLift ''U
+
+-- | "A"-side coin token
+data A = A
+PlutusTx.makeIsDataIndexed ''A [('A, 0)]
+PlutusTx.makeLift ''A
+
+-- | "B"-side coin token
+data B = B
+PlutusTx.makeIsDataIndexed ''B [('B, 0)]
+PlutusTx.makeLift ''B
+
+-- | Pool-state coin token
+data PoolState = PoolState
+PlutusTx.makeIsDataIndexed ''PoolState [('PoolState, 0)]
+PlutusTx.makeLift ''PoolState
+
+-- | Liquidity-state coin token
+data Liquidity = Liquidity
+PlutusTx.makeIsDataIndexed ''Liquidity [('Liquidity, 0)]
+PlutusTx.makeLift ''Liquidity
+
+-- Note: An orphan instance here because of the newtype wrapper below.
+deriving anyclass instance ToSchema AssetClass
+
+-- | A single 'AssetClass'. Because we use three coins, we use a phantom type to track
+-- which one is which.
+newtype Coin a = Coin { unCoin :: AssetClass }
+  deriving stock   (Show, Generic)
+  deriving newtype (ToJSON, FromJSON, ToSchema, Eq, Prelude.Eq, Prelude.Ord)
+PlutusTx.makeIsDataIndexed ''Coin [('Coin, 0)]
+PlutusTx.makeLift ''Coin
+
+-- | Likewise for 'Integer'; the corresponding amount we have of the
+-- particular 'Coin'.
+newtype Amount a = Amount { unAmount :: Integer }
+  deriving stock   (Show, Generic)
+  deriving newtype (ToJSON, FromJSON, ToSchema, Eq, Ord, PrintfArg)
+  deriving newtype (Prelude.Eq, Prelude.Ord, Prelude.Num)
+  deriving newtype (AdditiveGroup, AdditiveMonoid, AdditiveSemigroup, MultiplicativeSemigroup)
+PlutusTx.makeIsDataIndexed ''Amount [('Amount, 0)]
+PlutusTx.makeLift ''Amount
+
+{-# INLINABLE coin #-}
+coin :: Coin a -> Amount a -> Value
+coin c a = assetClassValue (unCoin c) (unAmount a)
+
+{-# INLINABLE coinValueOf #-}
+coinValueOf :: Value -> Coin a -> Amount a
+coinValueOf v = Amount . assetClassValueOf v . unCoin
+
+{-# INLINABLE mkCoin #-}
+mkCoin:: CurrencySymbol -> TokenName -> Coin a
+mkCoin c = Coin . assetClass c
+
+newtype Uniswap = Uniswap
+    { usCoin :: Coin U
+    } deriving stock    (Show, Generic)
+      deriving anyclass (ToJSON, FromJSON, ToSchema)
+      deriving newtype  (Prelude.Eq, Prelude.Ord)
+PlutusTx.makeIsDataIndexed ''Uniswap [('Uniswap, 0)]
+PlutusTx.makeLift ''Uniswap
+
+data LiquidityPool = LiquidityPool
+    { lpCoinA :: Coin A
+    , lpCoinB :: Coin B
+    }
+    deriving (Show, Generic, ToJSON, FromJSON, ToSchema)
+PlutusTx.makeIsDataIndexed ''LiquidityPool [('LiquidityPool, 0)]
+PlutusTx.makeLift ''LiquidityPool
+
+instance Eq LiquidityPool where
+    {-# INLINABLE (==) #-}
+    x == y = (lpCoinA x == lpCoinA y && lpCoinB x == lpCoinB y) ||
+              -- Make sure the underlying coins aren't equal.
+             (unCoin (lpCoinA x) == unCoin (lpCoinB y) && unCoin (lpCoinB x) == unCoin (lpCoinA y))
+
+data UniswapAction = Create LiquidityPool | Close | Swap | Remove | Add
+    deriving Show
+PlutusTx.makeIsDataIndexed ''UniswapAction [ ('Create , 0)
+                                           , ('Close,   1)
+                                           , ('Swap,    2)
+                                           , ('Remove,  3)
+                                           , ('Add,     4)
+                                           ]
+PlutusTx.makeLift ''UniswapAction
+
+data UniswapDatum =
+      Factory [LiquidityPool]
+    | Pool LiquidityPool (Amount Liquidity)
+    deriving stock (Show)
+PlutusTx.makeIsDataIndexed ''UniswapDatum [ ('Factory, 0)
+                                          , ('Pool,    1)
+                                          ]
+PlutusTx.makeLift ''UniswapDatum
